+++ R/spread.R
@@ -42,6 +42,13 @@ spread <- function(data, key, value, fill = NA, convert = FALSE, drop = TRUE) {
 #' @export
 spread_ <- function(data, key_col, value_col, fill = NA, convert = FALSE,
                     drop = TRUE) {
+  if (!(key_col %in% names(data))) {
+    stop("Key column '", key_col, "' does not exist in input.", call. = FALSE)
+  }
+  if (!(value_col %in% names(data))) {
+    stop("Value column '", value_col, "' does not exist in input.", call. = FALSE)
+  }
+
   UseMethod("spread_")
 }